moved TransactionType and Transaction to same file
